# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Work In Progress",
    "description": "Dokumentasi API untuk proses inspeksi, SWA, dan validasi",
    "version": "1.0.0"
  },
  "paths": {
    "/api/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["Username", "Password", "UserGroup", "Name"],
                "properties": {
                  "Username": {
                    "type": "string",
                    "example": "john.doe"
                  },
                  "Password": {
                    "type": "string",
                    "example": "secret123"
                  },
                  "UserGroup": {
                    "type": "string",
                    "example": "admin"
                  },
                  "Name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "IsActivated": {
                    "type": "boolean",
                    "example": "false"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User created successfully"
                    },
                    "user": {
                      "properties": {
                        "UserId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "Username": {
                          "type": "string",
                          "example": "john.doe"
                        },
                        "UserGroup": {
                          "type": "string",
                          "example": "admin"
                        },
                        "IsActivated": {
                          "type": "boolean",
                          "example": true
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user dengan username dan password",
        "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["Username", "Password"],
                "properties": {
                  "Username": {
                    "type": "string",
                    "example": "bhuedies"
                  },
                  "Password": {
                    "type": "string",
                    "example": "secret123"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout pengguna",
        "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
        "responses": {
          "200": {
            "description": "Berhasil logout",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Berhasil logout"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get authenticated user",
        "operationId": "b5de204f5e452ef51327c691aa99e77c",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "UserId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "Username": {
                      "type": "string"
                    },
                    "UserGroup": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh token JWT",
        "operationId": "e7ab7446393758bb71fd1946a349e500",
        "responses": {
          "200": {
            "description": "Token berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Use a bearer token to access this API",
        "name": "Authorization",
        "in": "header",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}